+++ src/models/filters/CustomFilter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace wsydney76\contentoverview\models\filters;
+
+use wsydney76\contentoverview\models\Filter;
+
+class CustomFilter extends Filter
+{
+
+}